requirements.txt
diffusers 
transformers 
accelerate 
torch 
pillow